Move IRLS directive to seperate file.
