+++ examples/01-maps/plot_sumMap.py
@@ -14,6 +14,8 @@ model.
 """
 from discretize import TensorMesh
 from discretize.utils import active_from_xyz
+
+import SimPEG.directives._regularization
 from SimPEG import (
     utils,
     maps,
@@ -167,7 +169,7 @@ def run(plotIt=True):
     # Here is where the norms are applied
     # Use pick a threshold parameter empirically based on the distribution of
     #  model parameters
-    IRLS = directives.Update_IRLS(f_min_change=1e-3, minGNiter=1)
+    IRLS = SimPEG.directives._regularization.Update_IRLS(f_min_change=1e-3, minGNiter=1)
     update_Jacobi = directives.UpdatePreconditioner()
     inv = inversion.BaseInversion(invProb, directiveList=[IRLS, betaest, update_Jacobi])
