+++ examples/03-magnetics/plot_inv_mag_MVI_VectorAmplitude.py
@@ -9,7 +9,7 @@ The inverse problem uses the :class:'SimPEG.regularization.VectorAmplitude'
 regularization borrowed from ...
 
 """
-
+import SimPEG.directives._regularization
 from SimPEG import (
     data,
     data_misfit,
@@ -178,7 +178,9 @@ betaest = directives.BetaEstimate_ByEig(beta0_ratio=1e1)
 sensitivity_weights = directives.UpdateSensitivityWeights()
 
 # Here is where the norms are applied
-IRLS = directives.Update_IRLS(f_min_change=1e-3, max_irls_iterations=10, beta_tol=5e-1)
+IRLS = SimPEG.directives._regularization.Update_IRLS(
+    f_min_change=1e-3, max_irls_iterations=10, beta_tol=5e-1
+)
 
 # Pre-conditioner
 update_Jacobi = directives.UpdatePreconditioner()
