+++ examples/03-magnetics/plot_inv_mag_nonLinear_Amplitude.py
@@ -20,6 +20,8 @@ assumption in order to improve the recovery of a compact prism.
 import scipy as sp
 import numpy as np
 import matplotlib.pyplot as plt
+
+import SimPEG.directives._regularization
 from SimPEG import (
     data,
     data_misfit,
@@ -248,7 +250,7 @@ betaest = directives.BetaEstimate_ByEig(beta0_ratio=2)
 
 # Target misfit to stop the inversion,
 # try to fit as much as possible of the signal, we don't want to lose anything
-IRLS = directives.Update_IRLS(
+IRLS = SimPEG.directives._regularization.Update_IRLS(
     f_min_change=1e-3, minGNiter=1, beta_tol=1e-1, max_irls_iterations=5
 )
 update_Jacobi = directives.UpdatePreconditioner()
@@ -353,7 +355,7 @@ invProb = inverse_problem.BaseInvProblem(dmis, reg, opt)
 betaest = directives.BetaEstimate_ByEig(beta0_ratio=1)
 
 # Specify the sparse norms
-IRLS = directives.Update_IRLS(
+IRLS = SimPEG.directives._regularization.Update_IRLS(
     max_irls_iterations=10,
     f_min_change=1e-3,
     minGNiter=1,
