+++ tests/dask/test_grav_inversion_linear.py
@@ -4,6 +4,7 @@ import discretize
 from discretize.utils import active_from_xyz
 import dask
 import SimPEG.dask  # noqa: F401
+import SimPEG.directives._regularization
 from SimPEG import (
     utils,
     maps,
@@ -103,7 +104,9 @@ class GravInvLinProblemTest(unittest.TestCase):
         invProb = inverse_problem.BaseInvProblem(dmis, reg, opt, beta=1e2)
 
         # Here is where the norms are applied
-        IRLS = directives.Update_IRLS(max_irls_iterations=20, chifact_start=2.0)
+        IRLS = SimPEG.directives._regularization.Update_IRLS(
+            max_irls_iterations=20, chifact_start=2.0
+        )
         update_Jacobi = directives.UpdatePreconditioner()
         sensitivity_weights = directives.UpdateSensitivityWeights(everyIter=False)
         self.inv = inversion.BaseInversion(
