+++ tests/pf/test_grav_inversion_linear.py
@@ -3,6 +3,8 @@ import numpy as np
 
 import discretize
 from discretize.utils import active_from_xyz
+
+import SimPEG.directives._regularization
 from SimPEG import (
     utils,
     maps,
@@ -104,7 +106,7 @@ def test_gravity_inversion_linear(engine):
 
     # Here is where the norms are applied
     starting_beta = directives.BetaEstimateMaxDerivative(10.0)
-    IRLS = directives.Update_IRLS()
+    IRLS = SimPEG.directives._regularization.Update_IRLS()
     update_Jacobi = directives.UpdatePreconditioner()
     sensitivity_weights = directives.UpdateSensitivityWeights(every_iteration=False)
     inv = inversion.BaseInversion(
