+++ tests/pf/test_mag_inversion_linear.py
@@ -1,6 +1,8 @@
 import unittest
 import discretize
 from discretize.utils import active_from_xyz
+
+import SimPEG.directives._regularization
 from SimPEG import (
     utils,
     maps,
@@ -112,7 +114,9 @@ class MagInvLinProblemTest(unittest.TestCase):
         betaest = directives.BetaEstimate_ByEig()
 
         # Here is where the norms are applied
-        IRLS = directives.Update_IRLS(f_min_change=1e-4, minGNiter=1)
+        IRLS = SimPEG.directives._regularization.Update_IRLS(
+            f_min_change=1e-4, minGNiter=1
+        )
         update_Jacobi = directives.UpdatePreconditioner()
         sensitivity_weights = directives.UpdateSensitivityWeights(everyIter=False)
         self.inv = inversion.BaseInversion(
