+++ tests/pf/test_mag_inversion_linear_Octree.py
@@ -3,6 +3,8 @@ import unittest
 import numpy as np
 
 from discretize.utils import mesh_builder_xyz, refine_tree_xyz, active_from_xyz
+
+import SimPEG.directives._regularization
 from SimPEG import (
     directives,
     maps,
@@ -129,7 +131,7 @@ class MagInvLinProblemTest(unittest.TestCase):
         )
 
         invProb = inverse_problem.BaseInvProblem(dmis, reg, opt, beta=1e6)
-        IRLS = directives.Update_IRLS()
+        IRLS = SimPEG.directives._regularization.Update_IRLS()
         update_Jacobi = directives.UpdatePreconditioner()
         sensitivity_weights = directives.UpdateSensitivityWeights()
         self.inv = inversion.BaseInversion(
