+++ tests/pf/test_mag_vector_amplitude.py
@@ -1,4 +1,6 @@
 import unittest
+
+import SimPEG.directives._regularization
 from SimPEG import (
     directives,
     maps,
@@ -130,7 +132,7 @@ class MVIProblemTest(unittest.TestCase):
         # Here is where the norms are applied
         # Use pick a treshold parameter empirically based on the distribution of
         #  model parameters
-        IRLS = directives.Update_IRLS(
+        IRLS = SimPEG.directives._regularization.Update_IRLS(
             f_min_change=1e-3, max_irls_iterations=10, beta_tol=5e-1
         )
