+++ tutorials/02-linear_inversion/plot_inv_2_inversion_irls.py
@@ -22,6 +22,7 @@ import matplotlib.pyplot as plt
 
 from discretize import TensorMesh
 
+import SimPEG.directives._regularization
 from SimPEG import (
     simulation,
     maps,
@@ -175,7 +176,9 @@ inv_prob = inverse_problem.BaseInvProblem(dmis, reg, opt)
 sensitivity_weights = directives.UpdateSensitivityWeights(everyIter=False)
 
 # Reach target misfit for L2 solution, then use IRLS until model stops changing.
-IRLS = directives.Update_IRLS(max_irls_iterations=40, minGNiter=1, f_min_change=1e-4)
+IRLS = SimPEG.directives._regularization.Update_IRLS(
+    max_irls_iterations=40, minGNiter=1, f_min_change=1e-4
+)
 
 # Defining a starting value for the trade-off parameter (beta) between the data
 # misfit and the regularization.
