+++ tutorials/04-magnetics/plot_inv_2a_magnetics_induced.py
@@ -34,6 +34,8 @@ import tarfile
 
 from discretize import TensorMesh
 from discretize.utils import active_from_xyz
+
+import SimPEG.directives._regularization
 from SimPEG.potential_fields import magnetics
 from SimPEG.utils import plot2Ddata, model_builder
 from SimPEG import (
@@ -293,7 +295,7 @@ save_iteration = directives.SaveOutputEveryIteration(save_txt=False)
 
 # Defines the directives for the IRLS regularization. This includes setting
 # the cooling schedule for the trade-off parameter.
-update_IRLS = directives.Update_IRLS(
+update_IRLS = SimPEG.directives._regularization.Update_IRLS(
     f_min_change=1e-4,
     max_irls_iterations=30,
     coolEpsFact=1.5,
