+++ tutorials/12-seismic/plot_inv_1_tomography_2D.py
@@ -29,6 +29,7 @@ import tarfile
 
 from discretize import TensorMesh
 
+import SimPEG.directives._regularization
 from SimPEG import (
     data,
     maps,
@@ -240,7 +241,7 @@ inv_prob = inverse_problem.BaseInvProblem(dmis, reg, opt)
 #
 
 # Reach target misfit for L2 solution, then use IRLS until model stops changing.
-update_IRLS = directives.Update_IRLS(
+update_IRLS = SimPEG.directives._regularization.Update_IRLS(
     f_min_change=1e-4,
     max_irls_iterations=30,
     coolEpsFact=1.5,
