+++ tutorials/_temporary/plot_inv_1_em1dtm_stitched_skytem.py
@@ -20,6 +20,7 @@ from matplotlib import pyplot as plt
 from discretize import TensorMesh
 from pymatsolver import PardisoSolver
 
+import SimPEG.directives._regularization
 from SimPEG.utils import mkvc
 from SimPEG import (
     maps,
@@ -326,7 +327,7 @@ update_Jacobi = directives.UpdatePreconditioner()
 save_iteration = directives.SaveOutputEveryIteration(save_txt=False)
 
 
-update_IRLS = directives.Update_IRLS(
+update_IRLS = SimPEG.directives._regularization.Update_IRLS(
     max_irls_iterations=20,
     minGNiter=1,
     fix_Jmatrix=True,
